fix(setup): add missing WiFi.h include + signed/unsigned cast
PIO build failed because main.cpp calls WiFi.status() == WL_CONNECTED
in the setup completion check but doesn't include WiFi.h directly.
Also fix signed-to-unsigned conversion warning in dns_respond call.

+++ firmware/canary/lib/securacv_setup/src/securacv_setup.cpp
@@ -163,7 +163,7 @@ void setup_dns_process(void) {
   int len = s_dns_udp.read(buf, sizeof(buf));
   if (len < 12) return;
 
-  dns_respond(buf, len, s_dns_udp.remoteIP(), s_dns_udp.remotePort());
+  dns_respond(buf, (size_t)len, s_dns_udp.remoteIP(), s_dns_udp.remotePort());
 }
 
 bool setup_start_captive_portal(void) {
